# internal/api/spec/openapi.yaml
openapi: 3.1.0
info:
  title: pool.biswas.me
  version: "1.0"
  summary: Pool water chemistry, treatment planning and cost-of-ownership tracking.
  description: |
    Every action the web interface performs is one of these calls — there is no
    separate API surface to keep in sync.

    **Authentication.** Send a personal API key as `Authorization: Bearer
    pool_sk_…`, or as `X-API-Key`. Create one under Settings → API keys in the
    app; the plaintext is shown once and cannot be recovered. A key carries a
    `read` or `read,write` scope, enforced on the HTTP method: a read key
    cannot POST, PATCH or DELETE. Browser sessions authenticate the same
    endpoints with a cookie.

    **Chemistry.** Readings are stored as given. Combined chlorine, adjusted
    alkalinity, the Langelier saturation index and the 0–100 water quality
    score are computed server-side from the pool's volume, surface and
    sanitizer, so a test posted through the API is scored identically to one
    typed into the app.

    **For an agent doing the reading.** The photo path is a convenience, not the
    only way in. An agent with a stronger model can do the whole job through
    the ordinary endpoints: `POST /api/tests` with the readings it transcribed,
    `POST /api/attachments` with the photo and that `test_id`, then
    `GET /api/tests/{id}/context` for the brief and `POST
    /api/tests/{id}/insight` with the analysis it wrote. Nothing about that
...
    say what did the reading.

...

...
servers:
...
security:
  - bearerAuth: []
  - apiKeyHeader: []
...
paths:
...
  /api/me:
    get:
      tags: [Account]
      summary: The authenticated account
      responses:
        "200":
          description: The account, its sign-in providers, and its AI configuration.
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: "#/components/schemas/User" }
                  providers: { type: object, additionalProperties: true }
                  ai:
                    type: object
                    properties:
                      configured: { type: boolean }
                      using_own_key: { type: boolean }
                      model: { type: string }
                      base_url: { type: string }
        "401": { $ref: "#/components/responses/Unauthorized" }
    patch:
      tags: [Account]
      summary: Change your name or password
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                password: { type: string, minLength: 8 }
      responses:
        "200":
          description: The updated account.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/me/ai:
    put:
      tags: [Account, AI]
      summary: Set your own LLM credentials
      description: |
        Your own key, endpoint and model, used in place of the server's so
        usage is billed to you. Any OpenAI-compatible endpoint works. Send an
        empty `api_key` to go back to the server's.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                api_key: { type: string }
                base_url: { type: string, examples: ["https://integrate.api.nvidia.com/v1"] }
                model: { type: string, examples: ["meta/llama-3.2-90b-vision-instruct"] }
      responses:
        "200": { description: Saved. }
...
  /api/me/mfa/totp/begin:
    post:
      tags: [Account]
      summary: Start two-factor setup
      description: |
        Issues a secret and leaves it unconfirmed. The factor is not on until a
        code computed from it comes back, so a setup abandoned halfway cannot
        lock anybody out of their own account.
      responses:
        "200":
          description: The secret, the otpauth URI, and where to fetch the QR.
          content:
            application/json:
              schema:
                type: object
                properties:
                  secret: { type: string, description: Grouped for reading off a screen. }
                  uri: { type: string, examples: ["otpauth://totp/..."] }
                  qr_url: { type: string }
                  digits: { type: integer }
                  period_s: { type: integer }
        "409": { description: Already on for this account. }
  /api/me/mfa/totp/qr:
    get:
      tags: [Account]
      summary: The enrolment QR
      description: |
        A PNG, served from an authenticated endpoint rather than embedded in
        the page, so the secret does not end up in page source, browser history
        or a screenshot of the URL bar. Gone once enrolment is confirmed.
      responses:
        "200":
          description: The code.
          content:
            image/png:
              schema: { type: string, format: binary }
        "404": { description: No enrolment in progress. }
  /api/me/mfa/totp/confirm:
    post:
      tags: [Account]
      summary: Prove the secret arrived and switch the factor on
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code: { type: string, examples: ["123456"] }
      responses:
        "200":
          description: |
            Enabled, with the recovery codes. They are shown once and never
            again — they are the way back in when the phone is gone.
          content:
            application/json:
              schema:
                type: object
                properties:
                  enabled: { type: boolean }
                  recovery_codes: { type: array, items: { type: string } }
                  message: { type: string }
        "400": { description: The code did not match. }
  /api/me/mfa/totp/disable:
    post:
      tags: [Account]
      summary: Turn the second factor off
      description: |
        Requires the password again. Without that, anybody who found an
        unlocked browser could quietly remove the protection. The recovery
        codes go with it.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password: { type: string }
      responses:
        "200": { description: Off. }
        "401": { description: The password was wrong. }
  /api/me/mfa/recovery-codes:
    post:
      tags: [Account]
      summary: Issue a fresh set of recovery codes
      description: Replaces the current set, which stops working immediately. Requires the password.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password: { type: string }
      responses:
        "200":
          description: The new codes, shown once.
          content:
            application/json:
              schema:
                type: object
                properties:
                  recovery_codes: { type: array, items: { type: string } }
...
  /api/me/passkeys/finish:
    post:
      tags: [Account]
      summary: Record the new passkey
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [challenge, response]
              properties:
                challenge: { type: string }
                name: { type: string, description: A label, defaulted from the device. }
                response: { type: object, additionalProperties: true }
      responses:
        "201":
          description: The registered passkey.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Passkey" }
        "400": { description: The attestation did not verify. }
        "409": { description: Already registered. }
  /api/me/passkeys/{id}:
    parameters: [{ $ref: "#/components/parameters/Id" }]
    patch:
      tags: [Account]
      summary: Rename a passkey
      responses:
        "200": { description: Renamed. }
    delete:
      tags: [Account]
      summary: Remove a passkey
      responses:
        "200": { description: Removed. }
  /api/auth/mfa:
    post:
      tags: [Account]
      summary: Finish a sign-in that stopped at the second factor
      description: |
        `POST /api/auth/login` answers `mfa_required` with a challenge instead
        of a session when the account has a second factor. Send the six digits
        here, or one of the recovery codes — each of those works once.

        The challenge is single use and expires in five minutes, so a
        half-finished login left on a shared machine is not a standing
        invitation.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [challenge, code]
              properties:
                challenge: { type: string }
                code: { type: string }
      responses:
        "200":
          description: Signed in.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401": { description: Wrong code, or the challenge has been spent. }
...
  /api/auth/passkey/finish:
    post:
      tags: [Account]
      summary: Complete a passwordless sign-in
      description: |
        A passkey is already two factors — something you have, unlocked by
        something you are or know — so it signs in on its own without the code.
        An authenticator whose signature counter has gone backwards is refused:
        that is what a cloned key looks like.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [challenge, response]
              properties:
                challenge: { type: string }
                response: { type: object, additionalProperties: true }
      responses:
        "200":
          description: Signed in.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401": { description: The assertion was not accepted. }
  /api/me/ai/providers:
    get:
      tags: [Account, AI]
      summary: Your own model chains
      description: |
        Two chains — `text` for the analysis, `vision` for reading a
        photographed test sheet — of up to three slots each. Slot 1 is tried
        first and each one below it is a fallback, so a provider that is
        rate-limiting or down moves to the next rather than failing.

        Keys are never returned. `has_key` and `key_hint` say whether a slot
        has one and show the last four characters, which is how the vendors
        print back a key you already own.

        An account with any slot configured uses only its own providers. The
        operator's key serves everyone else only when the instance sets
        `POOL_AI_SHARED=true`; otherwise it belongs to the operator alone.
      responses:
        "200":
          description: The configured chains, and the providers on offer.
          content:
            application/json:
              schema:
                type: object
                properties:
                  text:
                    type: array
                    items: { $ref: "#/components/schemas/AIProvider" }
                  vision:
                    type: array
                    items: { $ref: "#/components/schemas/AIProvider" }
                  using_own_keys: { type: boolean }
                  server_chain:
                    type: array
                    items: { type: string }
                    description: The operator's chain, named but not usable unless shared.
                  max_slots: { type: integer }
                  providers:
                    type: array
                    description: Providers the interface offers, with a model known to work.
                    items:
                      type: object
                      properties:
                        name: { type: string }
                        label: { type: string }
                        base_url: { type: string }
                        suggested_model: { type: string }
                        vision: { type: boolean }
                        free: { type: boolean }
                        sign_up: { type: string }
    put:
      tags: [Account, AI]
      summary: Set one slot of one chain
      description: |
        Leave `api_key` empty on a slot that already has one and the stored key
        is kept — so a model or an endpoint can be changed without the secret
        making another round trip through the browser.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [slot, provider]
              properties:
                kind: { type: string, enum: [text, vision], default: text }
                slot: { type: integer, minimum: 1, maximum: 3 }
                provider: { type: string, examples: [deepseek, nvidia, anthropic, openai] }
                model: { type: string }
                base_url:
                  type: string
                  description: Only for an endpoint go-ai does not already know.
                api_key: { type: string }
      responses:
        "200":
          description: The saved slot, key excluded.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AIProvider" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/me/ai/providers/{kind}/{slot}:
    delete:
      tags: [Account, AI]
      summary: Remove one slot
      description: |
        A gap is left rather than renumbering: the chain is read in slot order,
        and silently promoting a backup is not something to do behind
        somebody's back.
      parameters:
        - { name: kind, in: path, required: true, schema: { type: string, enum: [text, vision] } }
        - { name: slot, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: Removed. }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/keys/{id}:
    delete:
      tags: [Account]
      summary: Revoke an API key
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        "200": { description: Revoked. }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/pools:
    get:
      tags: [Pools]
      summary: List your pools
      responses:
        "200":
          description: Every pool on the account.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Pool" }
    post:
      tags: [Pools]
      summary: Create a pool
      description: |
        `volume_l` drives every dose calculation, and `surface` and `sanitizer`
        decide which ideal ranges a reading is scored against — a vinyl salt
        pool and a plaster chlorine pool are graded differently.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PoolInput" }
      responses:
        "201":
          description: The created pool.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Pool" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409": { description: A pool of that name already exists. }
  /api/pools/{id}:
    parameters: [{ $ref: "#/components/parameters/Id" }]
    get:
      tags: [Pools]
      summary: Fetch a pool
      responses:
        "200":
          description: The pool.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Pool" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [Pools]
      summary: Update a pool
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PoolInput" }
      responses:
        "200":
          description: The updated pool.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Pool" }
    delete:
      tags: [Pools]
      summary: Delete a pool
      description: Takes its tests, logbook entries, seasons and attachments with it.
      responses:
        "200": { description: Deleted. }
...
  /api/companies/{id}:
    parameters: [{ $ref: "#/components/parameters/Id" }]
    patch:
      tags: [Pools]
      summary: Update a company
      responses:
        "200": { description: The updated company. }
    delete:
      tags: [Pools]
      summary: Delete a company
      responses:
        "200": { description: Deleted. }
...
  /api/treatments/{id}/applied:
    post:
      tags: [Water tests]
      summary: Mark a dose as applied
      parameters: [{ $ref: "#/components/parameters/Id" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                applied: { type: boolean }
      responses:
        "200": { description: The updated treatment. }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: "`Authorization: Bearer pool_sk_…`"
    apiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
...